refactor(consensus): rename papyrus_node/main to run and move the main to its own small file
This is the overall pattern we are looking to follow, small mains that can just plug in their
changes to the general run framework.
